Move Status components in folder
