+++ front/app/containers/InitiativesShow/ReactionControl/Status/Ineligible.tsx
@@ -8,13 +8,13 @@ import {
   Icon,
   IconTooltip,
 } from '@citizenlab/cl2-component-library';
-import { StatusWrapper, StatusExplanation } from './SharedStyles';
-import ProposalProgressBar from './ProposalProgressBar';
+import { StatusWrapper, StatusExplanation } from '../SharedStyles';
+import ProposalProgressBar from '../ProposalProgressBar';
 import Button from 'components/UI/Button';
 import T from 'components/T';
-import messages from './messages';
+import messages from '../messages';
 import { FormattedMessage } from 'utils/cl-intl';
-import { StatusComponentProps } from './Status';
+import { StatusComponentProps } from '.';
 
 const StatusIcon = styled(Icon)`
   path {
