+++ front/app/containers/InitiativesShow/ReactionControl/Status/ReviewPending.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { colors, Box, Icon } from '@citizenlab/cl2-component-library';
-import { StatusWrapper, StatusExplanation } from './SharedStyles';
+import { StatusWrapper, StatusExplanation } from '../SharedStyles';
 import T from 'components/T';
-import messages from './messages';
+import messages from '../messages';
 import { FormattedMessage } from 'utils/cl-intl';
-import { StatusComponentProps } from './Status';
+import { StatusComponentProps } from '.';
 
 const ReviewPending = ({ initiativeStatus }: StatusComponentProps) => {
   return (
